refactor(governance): split with-governance into settlement, submission, and recovery modules
with-governance.ts had grown to 1,171 lines, well past the constitutional
500-line cap. This splits the activation runtime into cohesive modules
wired through an explicit GovernanceRuntimeDeps object, keeping
withGovernance() and activateGovernanceRuntime() as the composition root.
Behavior is unchanged; function bodies were moved verbatim.

Before: with-governance.ts 1,171 lines
After:
- with-governance.ts        242 lines (composition root)
- settlement.ts             325 lines (finalizeApprovedExecution, compensation, terminal-outcome helpers)
- submission.ts             341 lines (createSubmission, settleSubmission, authority decision, approve/reject)
- recovery.ts               143 lines (resumeStoredSettlement, resumePendingSettlements)
- settlement-observation.ts 335 lines (waitForSettlement and settlement result projection)
- runtime-deps.ts            39 lines (shared GovernanceRuntimeDeps interface)

Public API (src/index.ts) is unchanged; all governance (83) and
activation-cts (30) tests pass with unchanged counts.
